uTurnOff now uses "reboot" and "shutdown -r" commands for reboot.

+++ xpde/src/core/xpwm/uTurnOff.pas
@@ -75,6 +75,7 @@ end;
 
 procedure TTurnoff.restartClick(Sender: TObject);
 begin
+	if Libc.system('reboot')<>-1 then
         Libc.system('shutdown -t3 -r now');
 end;
